Non-functional: refactor utility threads for websocket tests into separate class.
